Create new WYSIWYG clone (WIP)
Blocks procedurally created from a const int array. TODO load from a file. (This will take some work because there is no working infrastructure for disabling vertex rendering yet.)

Player is controlled by the keyboard (though not legally and not accounting for direction

Block size is reduced by half so multiplying the model vertex coordinates by a hard-code 0.5 is no longer necessary. (the game mechanics are easiest when the block size is 1x1x1, but the Blender default cube is 2x2x2)