Further restructuring for GUI
A simple GUI was created and overlays on top of the rendered point cloud.
This GUI just has a few sliders for the camera position and direction for now.
The app has been moved to the app/ directory.
Plus all the sliders can be softened.

Right now, only a fraction of the points in the cloud are shown in case large point clouds are displayed because some graphics cards (like my GTX 1050M) are bottlenecked when rendering enough vertices. This can be changed in GUI.cpp, line 172. Someday, display all the points until the frame rate is below a threshold, then keep reducing the number of visible points until the frame rate is sufficient.